split router types so we don't need to rerun all documentation when doing some change on them
