+++ packages/router/examples/full-example.app.ts
@@ -1,7 +1,4 @@
-import {RpcError, StatusCodes} from '@mionkit/core';
-import {registerRoutes, initRouter, Route} from '@mionkit/router';
-import type {CallContext, HeaderHookDef, HookDef, RawHookDef, Routes} from '@mionkit/router';
-import type {APIGatewayEvent} from 'aws-lambda';
+import type {CallContext} from '@mionkit/router';
 
 export interface User {
     id: number;
