+++ packages/router/index.ts
@@ -5,7 +5,12 @@
  * The software is provided "as is", without warranty of any kind.
  * ######## */
 
-export * from './src/types';
+export * from './src/types/context';
+export * from './src/types/definitions';
+export * from './src/types/general';
+export * from './src/types/guards';
+export * from './src/types/handlers';
+export * from './src/types/remote';
 export * from './src/constants';
 export * from './src/router';
 export * from './src/dispatch';
