+++ website/content/2.docs/1.router/1.routes.md
@@ -80,7 +80,7 @@ Handler
 
 A Route can be a simple function `Handler` with the call context as first parameter. 
 
-<!-- embedme ../../../../packages/router/src/types.ts#L14-L19 -->
+<!-- embedme ../../../../packages/router/src/types/handlers.ts#L15-L20 -->
 ```ts
 export type Handler<Context extends CallContext = CallContext, Ret = any> = (
     /** Call Context */
@@ -97,7 +97,7 @@ RouteDef
 #code
 We can use a `RouteDef` object when we want to customize the default behavior of a route.
 
-<!-- embedme ../../../../packages/router/src/types.ts#L46-L55 -->
+<!-- embedme ../../../../packages/router/src/types/definitions.ts#L15-L24 -->
 ```ts
 export interface RouteDef<Context extends CallContext = CallContext, Ret = any> {
     /** description of the route, mostly for documentation purposes */
