+++ website/content/2.docs/1.router/2.hooks.md
@@ -147,7 +147,7 @@ const logger = {
 #name
 HookBase (HookOptions)
 #code
-<!-- embedme ../../../../packages/router/src/types.ts#L57-L66 -->
+<!-- embedme ../../../../packages/router/src/types/definitions.ts#L26-L35 -->
 ```ts
 interface HookBase {
     /** Executes the hook even if an error was thrown previously */
@@ -166,7 +166,7 @@ interface HookBase {
 #name
 HookDef
 #code
-<!-- embedme ../../../../packages/router/src/types.ts#L69-L72 -->
+<!-- embedme ../../../../packages/router/src/types/definitions.ts#L38-L41 -->
 ```ts
 export interface HookDef<Context extends CallContext = CallContext, Ret = any> extends HookBase {
     /** Hook handler */
@@ -179,7 +179,7 @@ export interface HookDef<Context extends CallContext = CallContext, Ret = any> e
 #name
 HeaderHookDef
 #code
-<!-- embedme ../../../../packages/router/src/types.ts#L75-L83 -->
+<!-- embedme ../../../../packages/router/src/types/definitions.ts#L44-L52 -->
 ```ts
 export interface HeaderHookDef<
     Context extends CallContext = CallContext,
@@ -197,7 +197,7 @@ export interface HeaderHookDef<
 #name
 RawHookDef
 #code
-<!-- embedme ../../../../packages/router/src/types.ts#L89-L97 -->
+<!-- embedme ../../../../packages/router/src/types/definitions.ts#L58-L66 -->
 ```ts
 export interface RawHookDef<
     Context extends CallContext = CallContext,
