+++ website/content/2.docs/1.router/3.call-context.md
@@ -96,7 +96,7 @@ The [`Request`](./request-and-response#request) and [`Response`](./request-and-r
 those can be accessed only using a [`RawHook`](./hooks#raw-hooks) and depends on the environment (server or serverless).
 
 
-<!-- embedme ../../../../packages/router/src/types.ts#L200-L209 -->
+<!-- embedme ../../../../packages/router/src/types/context.ts#L14-L23 -->
 ```ts
 export type CallContext<SharedData = any> = {
     /** Route's path after internal transformation */
