+++ website/content/2.docs/1.router/5.request-and-response.md
@@ -67,7 +67,7 @@ MionRequest
 
 mion's Request object, does not depends on the underlay native request.
 
-<!-- embedme ../../../../packages/router/src/types.ts#L217-L226 -->
+<!-- embedme ../../../../packages/router/src/types/context.ts#L31-L40 -->
 ```ts
 export interface MionRequest {
     /** parsed headers */
@@ -90,7 +90,7 @@ MionResponse
 
 mion's Response object, does not depends on the underlay native response.
 
-<!-- embedme ../../../../packages/router/src/types.ts#L229-L240 -->
+<!-- embedme ../../../../packages/router/src/types/context.ts#L43-L54 -->
 ```ts
 export interface MionResponse {
     /** response http status code */
@@ -117,7 +117,7 @@ mion's headers object, does not depends on the underlay native response.
 
 Similar to the Headers [Fetch API](https://developer.mozilla.org/en-US/docs/Web/API/Headers){target="_blank"}
 
-<!-- embedme ../../../../packages/router/src/types.ts#L246-L255 -->
+<!-- embedme ../../../../packages/router/src/types/context.ts#L60-L69 -->
 ```ts
 export interface MionHeaders {
     append(name: string, value: HeaderValue): void;
