+++ website/content/2.docs/3.serverless/1.aws-lambda.md
@@ -100,7 +100,7 @@ RouterOptions
 
 Basic options to configure mion router. These options are independent of the environment mion is being used (server or serverless).
 
-<!-- embedme ../../../../packages/router/src/types.ts#L115-L138 -->
+<!-- embedme ../../../../packages/router/src/types/general.ts#L30-L53 -->
 ```ts
 export interface RouterOptions<Req = any, SharedData = any> extends CoreOptions {
     /** prefix for all routes, i.e: api/v1.
