+++ website/content/2.docs/4.client.md
@@ -220,7 +220,7 @@ RemoteMethodMetadata<myRoutes>
 
 Required Metadata from Hook and Routes required for validation and serialization in the client.
 
-<!-- embedme ../../../packages/router/src/types.ts#L315-L329 -->
+<!-- embedme ../../../packages/router/src/types/remote.ts#L68-L82 -->
 ```ts
 export interface RemoteMethodMetadata<H extends Handler = any> {
     /** Type reference to the route handler, it's runtime value is actually null, just used statically by typescript. */
